src/requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.3
gevent==1.0.2
googlemaps==2.4.4
Werkzeug==0.11.9
pyjwt==1.3.0
python-Levenshtein==0.12.0
python-resize-image==1.1.10
psycopg2==2.6.2
